backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
celery>=5.3
django-celery-beat>=2.5
django-celery-results>=2.5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.3
stripe>=7.0
python-dotenv>=1.0
redis>=5.0
Pillow>=10.0
django-crispy-forms>=2.1
django-extensions>=3.2
factory-boy>=3.3
Faker>=23.0
requests>=2.31
qrcode>=7.4
gunicorn>=21.2
whitenoise>=6.5
